python_grpc_lab/requirements.txt
grpcio==1.75.1
grpcio-tools==1.75.1
protobuf>=6.31.1,<7.0.0